# backend/swagger.yaml
openapi: 3.0.3
info:
  title: SaaS Multi-Tenant Backend Platform API
  description: |
    A comprehensive multi-tenant SaaS backend platform built with Flask.

    ## Features
    - Multi-tenant architecture with database isolation
    - JWT-based authentication with refresh tokens
    - File management with MD5 deduplication
    - Role-based access control (Admin, User, Viewer)
    - Document management system
    - Kafka integration (placeholder for Phase 6)

    ## Authentication
    Most endpoints require JWT authentication. Include the access token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```

    ## Rate Limiting
    API rate limiting will be implemented in Phase 6.

    ## Error Responses
    All endpoints return consistent error responses with the following structure:
    ```json
    {
      "success": false,
      "message": "Error description",
      "error": "ERROR_CODE"
    }
    ```
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:4999
    description: Development server
  - url: http://localhost:8080
    description: Docker development server
  - url: https://api.example.com
    description: Production server (Phase 7)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
  schemas:
    Error:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "An error occurred"
        error:
          type: string
          example: "ERROR_CODE"
...
    Document:
      type: object
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
          example: "report.pdf"
        mime_type:
          type: string
          example: "application/pdf"
        file_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        created_by:
          type: string
          format: uuid
        updated_at:
          type: string
          format: date-time
    File:
      type: object
      properties:
        id:
          type: string
          format: uuid
        md5_hash:
          type: string
          example: "5d41402abc4b2a76b9719d911017c592"
        s3_path:
          type: string
          example: "tenants/123/files/5d/41/..."
        file_size:
          type: integer
          example: 1048576
        created_at:
          type: string
          format: date-time
        document_count:
          type: integer
          example: 3
        is_orphaned:
          type: boolean
          example: false
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
        pages:
          type: integer
          example: 5
...
paths:
...
  /api/files/{tenant_id}/files:
    get:
      tags:
        - Files
      summary: List tenant files
      description: Returns paginated list of files with deduplication info
      security:
        - BearerAuth: []
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the tenant
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: include_stats
          in: query
          schema:
            type: boolean
            default: false
          description: Include document count and orphan status
      responses:
        '200':
          description: Files retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Files retrieved successfully"
                  data:
                    type: object
                    properties:
                      files:
                        type: array
                        items:
                          $ref: '#/components/schemas/File'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
                      stats:
                        type: object
                        properties:
                          total_files:
                            type: integer
                          total_storage_bytes:
                            type: integer
                          total_storage_mb:
                            type: number
  /api/files/{tenant_id}/files/{file_id}:
    get:
      tags:
        - Files
      summary: Get file details
      description: Returns detailed information about a specific file
      security:
        - BearerAuth: []
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the tenant
        - name: file_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the file
      responses:
        '200':
          description: File retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "File retrieved successfully"
                  data:
                    allOf:
                      - $ref: '#/components/schemas/File'
                      - type: object
                        properties:
                          documents:
                            type: array
                            items:
                              $ref: '#/components/schemas/Document'

    delete:
      tags:
        - Files
      summary: Delete orphaned file
      description: Deletes an orphaned file (admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the tenant
        - name: file_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the file
      responses:
        '200':
          description: Orphaned file deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Orphaned file deleted successfully"
                  data:
                    type: object
                    properties:
                      file_id:
                        type: string
                        format: uuid
                      md5_hash:
                        type: string
                      freed_storage_bytes:
                        type: integer
                      freed_storage_mb:
                        type: number
        '400':
          description: File is not orphaned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized (not admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...